Basic enemy spawning
Spawn point finder is inefficient and ineffective.. might replace with
a better system that pregenerates a list of possible spawn points and
then just checks them until it finds a valid one.
Worse is the perf for spawning an enemy itself, 3-5ms. Out of scope
to figure out how to improve this I think.